personal

+++ app/Http/Controllers/admin/RolController.php
@@ -54,38 +54,51 @@ class RolController extends Controller
 
     public function update(ValidacionRol $request, $id)
     {
-        if($request->añadir == "on" )
-            $request->request->add(['añadir' => 1]);
-        else
-            $request->request->add(['añadir' => 0]);
-        if($request->editar == "on")
-            $request->request->add(['editar' => 1]);   
-        else
-            $request->request->add(['editar' => 0]); 
-        if($request->eliminar == "on")
-            $request->request->add(['eliminar' => 1]);   
-        else
-            $request->request->add(['eliminar' => 0]);
-        //dd($request->all());
-        Rol::findOrFail($id)->update($request->all());
-        return redirect('admin/rol')->with('mensaje', 'Rol actualizado con exito');
-       
+        if($id==1){
+            return redirect('admin/rol')->with('mensajeerror', 'no puedes editar el rol de Administrador');
+        }
+        else{
+            if($request->añadir == "on" )
+                $request->request->add(['añadir' => 1]);
+            else
+                $request->request->add(['añadir' => 0]);
+            if($request->editar == "on")
+                $request->request->add(['editar' => 1]);   
+            else
+                $request->request->add(['editar' => 0]); 
+            if($request->eliminar == "on")
+                $request->request->add(['eliminar' => 1]);   
+            else
+                $request->request->add(['eliminar' => 0]);
+            //dd($request->all());
+            Rol::findOrFail($id)->update($request->all());
+            return redirect('admin/rol')->with('mensaje', 'Rol actualizado con exito');
+        }
     }
 
     public function destroy(Request $request, $id)
     {
+        $aux=3; 
         if ($request->ajax()) { 
-            try {
-                //Eliminar registro
-                Rol::destroy($id);
-                $aux=1;
-            } catch (\Illuminate\Database\QueryException $e) {
-                $aux=0;
-            }          
+            if($id!=1){//solo va entrar cuando no es admin
+                try {
+                    //Eliminar registro
+                    Rol::destroy($id);
+                    $aux=1;
+                } catch (\Illuminate\Database\QueryException $e) {
+                    $aux=2;
+                } 
+            }  
+                            
             if ($aux==1) {
                 return response()->json(['mensaje' => 'ok']);
             } else {
-                return response()->json(['mensaje' => 'ng']);
+                if($aux==2){
+                    return response()->json(['mensaje' => 'ng']);
+                }
+                else{
+                    return response()->json(['mensaje' => 'admin']);
+                }
             }
         } else {
             abort(404);
